Add files via upload
First time tried to add JS in HTML & CSS. Added a small function of onclick to make a to-do list. Whenever we will type something in the to do list.. and click on the add button, the list will be created below with a delete button infront of it. and when clicked on the delete button, the list will be deleted,